skills/k8s-installer/scripts/requirements.txt
# K8S-Installer Skill Dependencies
# Python 3.11+

# SSH 連線
paramiko>=3.0.0

# CLI 介面
click>=8.0.0

# YAML 設定檔
pyyaml>=6.0
